+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.4.0 (UNRELEASED)
+
+* Add `delay` option for additional files that are written outside the plugin lifecycle
+* Assume a default delay of 5000ms in rollup 0.60.0-0.62.0 as a workaround for changed behavior (fixes #2)
+
 ## v1.3.0 (2017-08-21)
 
 * Compatibility with rollup 0.48